worked on bldrone

+++ BL/Bo/Drone.cs
@@ -15,7 +15,7 @@ namespace IBL
             public int DroneId { get; set; }
             public string DroneModel { get; set; }
             public WeightCategories Weight { get; set; }
-            public int BatteryStatus { get; set; }
+            public double BatteryStatus { get; set; }
             public DroneStatuses DroneStatus { get; set; }
             public ParcelInTransfer DeliveryTransfer { get; set; }
             public Location DroneLocation { get; set; }
